fix(type): register module hooks types
