+++ docs/content/docs/7.advanced/5.hooks.md
@@ -13,11 +13,9 @@ It can be used to modify the raw content from a `file` before it is transformed
 or modify the transform options.
 
 ```ts
-import type { FileBeforeParseHook } from '@nuxt/content'
-
 export default defineNuxtConfig({
   hooks: {
-    'content:file:beforeParse'(ctx: FileBeforeParseHook) {
+    'content:file:beforeParse'(ctx) {
       // ...
     }
   }
@@ -29,11 +27,9 @@ export default defineNuxtConfig({
 This hook is called after the content is parsed and before it is saved to the database.
 
 ```ts
-import type { FileAfterParseHook } from '@nuxt/content'
-
 export default defineNuxtConfig({
   hooks: {
-    'content:file:afterParse'(ctx: FileAfterParseHook) {
+    'content:file:afterParse'(ctx) {
       // ...
     }
   }
