+++ src/types/index.ts
@@ -1,4 +1,5 @@
 export type * from './collection'
+export type * from './hooks'
 export type * from './module'
 export type * from './navigation'
 export type * from './surround'
